+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and what APIs have changed, if applicable.
 
 ## [Unreleased]
 
+## [29.19.10] - 2021-07-16
+- Add hooks for customizing documentation (OPTIONS) response.
+  - Documentation renderers now get the request headers and resource models available during rendering.
+
 ## [29.19.9] - 2021-07-15
 - Relax extension schema check to make '@extension' annotation is optional for 1-to-1 injections.
 - Update RestliRouter to allow "bq", "action" as query parameter name for finder, "q" as action parameter name for action
@@ -5004,7 +5008,8 @@ patch operations can re-use these classes for generating patch messages.
 
 ## [0.14.1]
 
-[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.19.9...master
+[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.19.10...master
+[29.19.10]: https://github.com/linkedin/rest.li/compare/v29.19.9...v29.19.10
 [29.19.9]: https://github.com/linkedin/rest.li/compare/v29.19.8...v29.19.9
 [29.19.8]: https://github.com/linkedin/rest.li/compare/v29.19.7...v29.19.8
 [29.19.7]: https://github.com/linkedin/rest.li/compare/v29.19.6...v29.19.7